requirements.txt
langchain
langchain-community
langchain-openai
langchain-huggingface
langchain_text_splitters
chromadb
streamlit
pandas
python-dotenv
sentence-transformers
fastapi
uvicorn[standard]
requests